First pass, need to fix db querry and set up some tests using users and predefined workouts
